easy-rider-backend/pyproject.toml
[tool.poetry]
name = "easy-rider-backend"
version = "0.1.0"
description = ""
authors = ["Viacheslav Bessonov <viacheslav.a.bessonov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.5"
djangorestframework = "^3.11.0"
django-environ = "^0.4.5"
drf-nested-routers = "^0.91"
drf-access-policy = "^0.5.1"
djangorestframework_jwt = "^1.11.0"
django-cors-headers = "^3.2.1"
django-filter = "^2.2.0"
djangorestframework-camel-case = "^1.1.2"
environ = "^1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
flake8 = "^3.7.9"
isort = "^4.3.21"
bumpversion = "^0.5.3"
nose = "^1.3.7"
parameterized = "^0.7.1"
coverage = "^5.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
